refactor(history): extract reusable history evaluation helpers
- Move reusable offline history evaluation helpers into research.history_evaluation
- Keep scripts.evaluate_history focused on CLI parsing and main wiring
- Split helper tests from CLI script tests
- Preserve existing single-entry, bulk, summary, and JSON output behavior

Refs #72
